feat: add chunk-based map system with dynamic loading

+++ src/CryBits.Client/Spawners/BloodSplatSpawner.cs
@@ -11,7 +11,7 @@ internal static class BloodSplatSpawner
 {
     private const int FrameSize = 32;
 
-    public static EntityId Spawn(World world, short tileX, short tileY)
+    public static EntityId Spawn(World world, int tileX, int tileY)
     {
         var frame = Random.Shared.Next(0, 3);
         var sourceRect = new Rectangle(frame * FrameSize, 0, FrameSize, FrameSize);
