+++ src/CryBits.Definitions/Maps/MapNpc.cs
@@ -8,6 +8,6 @@ public partial class MapNpc
     public Guid NpcId { get; set; }
     public byte Zone { get; set; }
     public bool Spawn { get; set; }
-    public byte X { get; set; }
-    public byte Y { get; set; }
+    public int X { get; set; }
+    public int Y { get; set; }
 }
