+++ src/CryBits.Simulation/Components/AttackHit.cs
@@ -3,4 +3,4 @@ using MemoryPack;
 namespace CryBits.Simulation.Components;
 
 [MemoryPackable]
-public sealed partial record class AttackHit(long? VictimId, byte VictimTileX = 0, byte VictimTileY = 0);
+public sealed partial record class AttackHit(long? VictimId, int VictimTileX = 0, int VictimTileY = 0);
