+++ src/CryBits.Simulation/Components/NpcState.cs
@@ -4,4 +4,4 @@ using MemoryPack;
 namespace CryBits.Simulation.Components;
 
 [MemoryPackable]
-public sealed partial record class NpcState(byte Index, Guid NpcDefId, EntityId? TargetId = null);
+public sealed partial record class NpcState(int Index, Guid NpcDefId, EntityId? TargetId = null);
