+++ src/CryBits.Simulation/Events/LootDroppedEvent.cs
@@ -1,3 +1,3 @@
 namespace CryBits.Simulation.Events;
 
-public sealed partial record class LootDroppedEvent(long TickNumber, Guid MapId, byte X, byte Y, Guid ItemId, short Amount, long DespawnTick) : SimEvent(TickNumber);
+public sealed partial record class LootDroppedEvent(long TickNumber, Guid MapId, int X, int Y, Guid ItemId, short Amount, long DespawnTick) : SimEvent(TickNumber);
